+++ src/arrayShape.test.js
@@ -1,5 +1,4 @@
 var arrayShape = require('../dist/arrayShape').default;
-var assertionContext = require('../dist/assertionContext').default;
 
 describe('arrayShape() function', () => {
     it ('Matches correct arrays', () => {
@@ -9,15 +8,15 @@ describe('arrayShape() function', () => {
 
         var shape = arrayShape('number');
 
-        expect(shape(arrA)).toBe(true);
-        expect(shape(arrB)).toBe(false);
-        expect(shape(arrC)).toBe(false);
+        expect(() => shape(arrA)).not.toThrow();
+        expect(() => shape(arrB)).toThrow();
+        expect(() => shape(arrC)).toThrow();
     });
 
     it ('Matches empty arrays', () => {
         var shape = arrayShape('number');
 
-        expect(shape([])).toBe(true);
+        expect(() => shape([])).not.toThrow();
     });
 
     it ('Can optionally disallow empty arrays', () => {
@@ -26,18 +25,7 @@ describe('arrayShape() function', () => {
 
         var shape = arrayShape('number', { disallowEmpty: true });
 
-        expect(shape(arrA)).toBe(true);
-        expect(shape(arrB)).toBe(false);
-    });
-
-    it ('Throws when used in an assertion', () => {
-        assertionContext.push(0);
-
-        const arrA = [1, 2];
-        const shape = arrayShape('string');
-
-        expect(() => shape(arrA)).toThrow();
-
-        assertionContext.clear();
+        expect(() => shape(arrA)).not.toThrow();
+        expect(() => shape(arrB)).toThrow();
     });
 });