+++ src/arrayShape.ts
@@ -1,8 +1,6 @@
-import { ShapeProperty, RealTypeOfProperty } from "./shapeTypes";
-import { matchesSpecifier } from "./hasShape";
-import assertionContext from "./assertionContext";
-import { assertionError } from "./utils";
-import { assertSpecifier } from "./assertShape";
+import { Shape, RealTypeOfShape, CompiledShape, ArrayShape } from "./shapeTypes";
+import assertShape from "./assertShape";
+import context from "./context";
 
 interface ArrayShapeOptions {
     /**
@@ -11,46 +9,32 @@ interface ArrayShapeOptions {
     disallowEmpty?: boolean
 }
 
-const fail = (message: string) => {
-    if (assertionContext.isAsserting()) {
-        throw assertionError(message);
-    }
-
-    return false;
-}
-
 /**
- * Creates a predicate function which checks whether a value is an array
+ * Creates an assertion function which checks whether a value is an array
  *     of the specified type.
  *
  * By default, empty arrays are valid. This can be overridden
  *     using `options.disallowEmpty`.
  */
-export default <T extends ShapeProperty>(type: T, options?: ArrayShapeOptions) => {
+export default <T extends Shape>(shape: T, options?: ArrayShapeOptions): CompiledShape<ArrayShape<T>> => {
     const {
         disallowEmpty = false
     } = options ?? {};
 
-    return (target: unknown): target is RealTypeOfProperty<T>[] => {
-        if (!Array.isArray(target)) {
-            return fail(`Value is not an array. Received: ${target}`);
-        }
-        else if (disallowEmpty && target.length < 1) {
-            return fail('Array cannot be empty.');
-        }
+    return (target: unknown) => {
+        context.runInLayer(() => {
+            if (!Array.isArray(target)) {
+                throw `Value is not an array. Received: ${target}`;
+            }
+            else if (disallowEmpty && target.length < 1) {
+                throw 'Array cannot be empty.';
+            }
 
-        if (assertionContext.isAsserting()) {
             target.forEach((v, i) => {
-                assertionContext.push(i);
-
-                assertSpecifier(v, type);
-
-                assertionContext.pop();
+                context.runInLayer(i, () => {
+                    assertShape(v, shape);
+                });
             });
-
-            return true;
-        }
-
-        return target.every(v => matchesSpecifier(v, type));
+        });
     };
 };