+++ src/hasShape.ts
@@ -1,35 +1,20 @@
-import {
-    Shape,
-    ShapeProperty,
-    RealTypeOfShape,
-    RealTypeOfProperty
-} from './shapeTypes';
-import assertShape, { assertSpecifier } from './assertShape';
+import { Shape, RealTypeOfShape } from './shapeTypes';
+import assertShape from './assertShape';
 
 /**
- * Checks whether the given value matches a specifier.
+ * Given a `target` value and a `shape`, the target is checked to ensure
+ *     it has the specified 'shape'.
  *
- * Only intended to be used internally.
- */
-export const matchesSpecifier = <T extends ShapeProperty>(value: unknown, specifier: T): value is RealTypeOfProperty<T> => {
-    try { assertSpecifier(value, specifier); }
-    catch { return false; }
-    return true;
-};
-
-/**
- * Given a `target` value and a `shape` object, the target is checked to ensure
- *     it has the specified 'shape'. The shape should be a dictionary of
- *     property keys to type specifiers.
- *
- * Type specifiers may be any of the following:
+ * A shape may be any of the following:
+ * - An object acting as a dictionary of other shapes,
+ *     e.g. `{ name: 'string', age: 'number }`
  * - A `typeof` string (`'boolean'`, `'number'`, `'string'`, etc.)
  * - One of the strings `'any'` or `'unknown'` to specify any type,
  *     with matching TS types.
- * - A class or constructor function
- * - A regular expression ('Regex')
- * - A predicate function
- * - Another shape
+ * - A regular expression ('Regex').
+ * - A predicate function.
+ * - An assertion function.
+ * - A shape created through the `shape()` or `arrayShape()` functions.
  */
 const hasShape = <T extends Shape>(target: unknown, shape: T): target is RealTypeOfShape<T> => {
     try { assertShape(target, shape); }