+++ src/index.ts
@@ -1,5 +1,8 @@
+import assertShape from './assertShape';
+import arrayShape from './arrayShape';
+
 export { default } from './hasShape';
 export { default as arrayShape } from './arrayShape';
 export { default as assertShape } from './assertShape';
-export { default as compileShape, compileShapeAssertion } from './compileShape';
+export { default as shape } from './shape';
 export * from './shapeTypes';